feat: move logger file from services to shared
