+++ backend/shared/go.mod
@@ -1,3 +1,11 @@
 module github.com/unwelcome/FrameWorkTask1/backend/shared
 
 go 1.25.1
+
+require github.com/rs/zerolog v1.35.1
+
+require (
+	github.com/mattn/go-colorable v0.1.14 // indirect
+	github.com/mattn/go-isatty v0.0.22 // indirect
+	golang.org/x/sys v0.43.0 // indirect
+)
